旧reducer削除＆utils切り出し
